Split cacheObjects in separate files
