+++ package.json
@@ -18,9 +18,9 @@
     "test:typescript": "typings-tester --dir typescript_test",
     "clean": "rm -rf dist es lib",
     "compile": "npm run clean && npm run compile:es && npm run compile:commonjs && npm run compile:umd",
-    "compile:es": "babel -d es/ src/",
-    "compile:commonjs": "cross-env NODE_ENV=commonjs babel -d lib/ src/",
-    "compile:umd": "cross-env NODE_ENV=umd babel -d dist/ src/",
+    "compile:es": "babel --ignore __tests__ -d es/ src/",
+    "compile:commonjs": "cross-env NODE_ENV=commonjs babel --ignore __tests__ -d lib/ src/",
+    "compile:umd": "cross-env NODE_ENV=umd babel --ignore __tests__ -d dist/ src/",
     "contrib:add": "all-contributors add",
     "contrib:generate": "all-contributors generate",
     "preversion": "npm test",
