+++ src/index.js
@@ -1,101 +1,5 @@
 import { createSelector } from 'reselect';
-
-// @TODO Move it into a separate file and test it
-export class FlatCacheObject {
-  constructor() {
-    this._cache = {};
-  }
-  set(key, selectorFn) {
-    this._cache[key] = selectorFn;
-  }
-  get(key) {
-    return this._cache[key];
-  }
-  remove(key) {
-    delete this._cache[key];
-  }
-  clear() {
-    this._cache = {};
-  }
-}
-
-export class FifoCacheObject {
-  constructor({ cacheSize } = {}) {
-    if (cacheSize === undefined) {
-      throw new Error('Missing the required property `cacheSize`.');
-    }
-    if (!Number.isInteger(cacheSize) || cacheSize <= 0) {
-      throw new Error('The `cacheSize` property must be a positive integer value.');
-    }
-    this._cache = {};
-    this._cacheOrdering = [];
-    this._cacheSize = cacheSize;
-  }
-  set(key, selectorFn) {
-    this._cache[key] = selectorFn;
-    this._cacheOrdering.push(key);
-
-    if (this._cacheOrdering.length > this._cacheSize) {
-      const earliest = this._cacheOrdering.shift();
-      delete this._cache[earliest];
-    }
-  }
-  get(key) {
-    return this._cache[key];
-  }
-  remove(key) {
-    const index = this._cacheOrdering.indexOf(key);
-
-    if (index > -1) {
-      this._cacheOrdering.splice(index, 1);
-      delete this._cache[key];
-    }
-  }
-  clear() {
-    this._cache = {};
-    this._cacheOrdering = [];
-  }
-}
-
-export class LruCacheObject {
-  constructor ({ cacheSize } = {}) {
-    if (cacheSize === undefined) {
-      throw new Error('Missing the required property `cacheSize`.');
-    }
-    if (!Number.isInteger(cacheSize) || cacheSize <= 0) {
-      throw new Error('The `cacheSize` property must be a positive integer value.');
-    }
-    this._cache = {};
-    this._cacheOrdering = [];
-    this._cacheSize = cacheSize;
-  }
-  set(key, selectorFn) {
-    this._cache[key] = selectorFn;
-    this._registerCacheHit(key);
-    if (this._cacheOrdering.length > this._cacheSize) {
-        const earliest = this._cacheOrdering.shift();
-        delete this._cache[earliest];
-    }
-  }
-  get(key) {
-    this._registerCacheHit(key);
-    return this._cache[key];
-  }
-  remove(key) {
-    delete this._cache[key];
-  }
-  clear() {
-    this._cache = {};
-    this._cacheOrdering = [];
-  }
-  _registerCacheHit(key) {
-    const index = this._cacheOrdering.indexOf(key);
-    if (index !== -1) {
-        this._cacheOrdering.splice(index, 1);
-    }
-    this._cacheOrdering.push(key);
-  }
-}
+import FlatCacheObject from './cache/FlatCacheObject';
 
 export default function createCachedSelector(...funcs) {
   const defaultCacheCreator = FlatCacheObject;
